requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Embeddings (lightweight, CPU-only)
sentence-transformers==2.3.1

# Text processing
spacy==3.7.2
yake==0.4.8
sympy==1.12

# HTTP client for external APIs (Groq/OpenRouter)
httpx==0.26.0
aiohttp==3.9.1

# Vector store
faiss-cpu==1.7.4
numpy>=1.24.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Type hints
typing-extensions>=4.0.0
